Windows Compat for v2.0.0 release
Forgot to commit this because it was on my windows vm and it was like 2 am, but there we go. This was already in the v2 release, so it's fine. Also removes the unneeded webbrowser import because I was trying to use it to try and open the stuff in the default linux terminal emulator but apparently thats just not possible?? anyways, fixed